+++ src/components/projectInfo/dropDown/dropDown.tsx
@@ -1,12 +1,14 @@
 import { useState } from 'react';
 
+import type { TProjectPath } from '@/types/scenario/scenario';
+
 import * as S from '@/components/projectInfo/dropDown/dropDown.style';
 
 import ArrowDown from '@/assets/icons/arrow_down.svg?react';
 import ArrowUp from '@/assets/icons/arrow_up.svg?react';
 
 interface IDropdownProps {
-  options: string[];
+  options: TProjectPath[];
   onSelect: (value: string) => void;
   placeholder?: string;
 }
@@ -27,8 +29,8 @@ export default function Dropdown({ options, onSelect, placeholder }: IDropdownPr
       </S.DropdownHeader>
       <S.DropdownList isOpen={isOpen}>
         {options.map((option, index) => (
-          <S.DropdownListItem key={index} onClick={() => handleSelect(option)}>
-            {option}
+          <S.DropdownListItem key={index} onClick={() => handleSelect(option.path)}>
+            {option.path}
           </S.DropdownListItem>
         ))}
       </S.DropdownList>
