+++ src/components/scenario/scenarioModal/scenarioModal.tsx
@@ -1,27 +1,14 @@
-import { useEffect, useMemo, useState } from 'react';
-import type { SubmitHandler } from 'react-hook-form';
-import { useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
+import { useState } from 'react';
 
 import type { TRequestCharacterScenarioResponse } from '@/types/scenario/scenario';
-import { QUERY_KEYS } from '@/constants/querykeys/queryKeys';
-
-import { createCharacterModalScehma } from '@/utils/validate';
-import { queryClient } from '@/apis/queryClient';
 
 import { useDispatch } from '@/hooks/common/useCustomRedux.ts';
 import useChangeScenarioInfo from '@/hooks/scenario/useChangeScenarioInfo.ts';
-import useGetScenarioModalInfo from '@/hooks/scenario/useGetScenarioInfo';
 
-import Button from '@/components/common/button/button';
-import Input from '@/components/common/input/input';
-import ValidataionMessage from '@/components/common/input/validationMessage';
-import Loading from '@/components/common/loading/loading';
 import Modal from '@/components/common/modal/modal';
-import Dropdown from '@/components/scenario/dropDown/dropDown';
-import * as S from '@/components/scenario/scenarioModal/scenarioModal.style';
+import ScenarioModalStep1 from '@/components/scenario/scenarioModal/scenarioModal/scenarioModalStep1';
+import ScenarioModalStep2 from '@/components/scenario/scenarioModal/scenarioModal/scenarioModalStep2';
 
-import DelCircle from '@/assets/icons/del_circle.svg?react';
 import { closeModal } from '@/slices/modalSlice.ts';
 
 type TScenarioProps = {
@@ -29,127 +16,15 @@ type TScenarioProps = {
   currentPage: number;
 };
 
-type TFormValues = {
-  projectId: number;
-  characterName: string;
-  characterDescription: string;
-  accessPage: string[];
-};
-
-type TItem = {
-  step: string;
-  actionDescription: string;
-};
-
 export default function ScenarioModal({ projectId, currentPage }: TScenarioProps) {
   const dispatch = useDispatch();
 
   const [modalStep, setModalStep] = useState<number>(1); // 모달 단계 상태 (1: 역할 선택, 2: 역할 확인)
-  const [selectedOptions, setSelectedOptions] = useState<string[]>([]); // 선택된 옵션
-  const [isSubmitted, setIsSubmitted] = useState<boolean>(false);
-  const [scenarioId, setScenarioId] = useState<number>();
-  const [characterId, setCharacterId] = useState<number>();
   const [characterData, setCharacterData] = useState<TRequestCharacterScenarioResponse>();
-  const [errorMessage, setErrorMessage] = useState('');
 
-  const { useGetAllPaths } = useGetScenarioModalInfo({ projectId, currentPage });
   const { usePostCharacter, usePatchCharacter } = useChangeScenarioInfo();
-
-  const { data: PathData } = useGetAllPaths;
   const { mutate: postCharacter, isPending: postCharacterPending } = usePostCharacter;
   const { mutate: patchCharacter, isPending: patchCharacterPending } = usePatchCharacter;
-
-  const options = useMemo(() => PathData?.result?.projectPaths ?? [], [PathData]);
-
-  const {
-    register,
-    handleSubmit,
-    formState: { isValid, errors, touchedFields },
-  } = useForm<TFormValues>({
-    mode: 'onChange',
-    resolver: zodResolver(createCharacterModalScehma),
-  });
-
-  useEffect(() => {
-    setErrorMessage('');
-  }, [selectedOptions]);
-
-  // 역할 생성 함수
-
-  const invalidateQueries = () => {
-    queryClient.invalidateQueries({
-      queryKey: QUERY_KEYS.GET_CHARACTER_LIST({ projectId, currentPage }),
-    });
-  };
-
-  const onSubmit: SubmitHandler<TFormValues> = async (submitData) => {
-    setErrorMessage('');
-    if (!isSubmitted) {
-      postCharacter(
-        {
-          projectId: Number(projectId),
-          characterName: submitData.characterName,
-          characterDescription: submitData.characterDescription,
-          accessPage: selectedOptions,
-        },
-        {
-          onSuccess: (data) => {
-            setErrorMessage('');
-            setCharacterData(data);
-            setIsSubmitted(true);
-            setCharacterId(data.result.characterId);
-            setScenarioId(data.result.scenarioId);
-            setModalStep(2);
-            invalidateQueries();
-          },
-          onError: (error) => {
-            setErrorMessage(error.response?.data?.detail || 'An error occurred');
-          },
-        },
-      );
-    } else {
-      if (scenarioId !== undefined && characterId !== undefined) {
-        patchCharacter(
-          {
-            projectId: Number(projectId),
-            characterName: submitData.characterName,
-            characterDescription: submitData.characterDescription,
-            accessPage: selectedOptions,
-            characterId: characterId,
-            scenarioId: scenarioId,
-          },
-          {
-            onSuccess: (data) => {
-              setErrorMessage('');
-              setCharacterData(data);
-              setIsSubmitted(true);
-              setModalStep(2);
-              invalidateQueries();
-            },
-            onError: (error) => {
-              setErrorMessage(error.response?.data?.detail || 'An error occurred');
-            },
-          },
-        );
-      } else {
-        setErrorMessage('An error occurred');
-      }
-    }
-  };
-
-  // 선택된 옵션 추가 함수
-  const handleSelect = (value: string) => {
-    if (!selectedOptions.includes(value)) {
-      setSelectedOptions((prev) => [...prev, value]);
-    }
-  };
-
-  // 선택된 옵션 제거 함수
-  const handleRemove = (value: string) => {
-    setSelectedOptions((prev) => prev.filter((option) => option !== value));
-  };
-
-  const parsedData = characterData?.result.scenarioDescription ? JSON.parse(characterData.result.scenarioDescription) : null;
 
   return (
     <Modal
@@ -162,101 +37,18 @@ export default function ScenarioModal({ projectId, currentPage }: TScenarioProps
       clickOutside={false}
     >
       {modalStep === 1 && (
-        // 역할 선택 단계
-        <form onSubmit={handleSubmit(onSubmit)}>
-          <S.ModalContainer>
-            {(postCharacterPending || patchCharacterPending) && (
-              <S.LoadingContainer2>
-                <Loading />
-              </S.LoadingContainer2>
-            )}
-            <div>
-              <S.description>Define users for the registered project.</S.description>
-              <S.description>QASTUDIO will create a suitable scenario for you.</S.description>
-            </div>
-
-            <S.InputWrapper>
-              <S.SubTitle>Name</S.SubTitle>
-              {errors.characterName?.message && touchedFields.characterName && (
-                <S.ValidationWrapper>
-                  <ValidataionMessage message={errors.characterName.message} isError={true} />
-                </S.ValidationWrapper>
-              )}
-              <Input type="normal" placeholder="Define the target character in one sentence." {...register('characterName')} />
-            </S.InputWrapper>
-            <S.InputWrapper>
-              <S.SubTitle>Description</S.SubTitle>
-              {errors.characterDescription?.message && touchedFields.characterDescription && (
-                <S.ValidationWrapper>
-                  <ValidataionMessage message={errors.characterDescription.message} isError={true} />
-                </S.ValidationWrapper>
-              )}
-              <Input placeholder="Explain the character's purpose for using the project." type="normal" {...register('characterDescription')} />
-            </S.InputWrapper>
-            <S.InputWrapper>
-              <S.SubTitle>Access page</S.SubTitle>
-
-              <Dropdown options={options} onSelect={handleSelect} placeholder="Select pages accessible to the character." />
-            </S.InputWrapper>
-
-            <S.TagContainer>
-              {selectedOptions.map((option) => (
-                <Button key={option} type="tag" color="mint" icon={<DelCircle />} iconPosition="right" onClick={() => handleRemove(option)}>
-                  {option}
-                </Button>
-              ))}
-            </S.TagContainer>
-            <S.ButtonContainer>
-              {errorMessage !== '' ? <ValidataionMessage message={errorMessage} isError={!!errorMessage} /> : <div />}
-              <Button
-                color="blue"
-                onClick={handleSubmit(onSubmit)}
-                disabled={selectedOptions.length < 1 || postCharacterPending || patchCharacterPending || !isValid}
-              >
-                Create
-              </Button>
-            </S.ButtonContainer>
-          </S.ModalContainer>
-        </form>
+        <ScenarioModalStep1
+          postCharacter={postCharacter}
+          patchCharacter={patchCharacter}
+          projectId={projectId}
+          currentPage={currentPage}
+          setCharacterData={setCharacterData}
+          setModalStep={setModalStep}
+          postCharacterPending={postCharacterPending}
+          patchCharacterPending={patchCharacterPending}
+        />
       )}
-      {modalStep == 2 && (
-        <S.ConfirmModalContainer>
-          <S.DescriptionContainer>
-            <S.DescriptionItem>
-              <S.SubTitle>Character</S.SubTitle>
-              <S.DescriptionContent>{characterData?.result.characterName}</S.DescriptionContent>
-            </S.DescriptionItem>
-            <S.DescriptionItem>
-              <S.SubTitle>Description</S.SubTitle>
-              <S.DescriptionContent>{characterData?.result.characterDescription}</S.DescriptionContent>
-            </S.DescriptionItem>
-            <S.DescriptionItem>
-              <S.SubTitle>Access control</S.SubTitle>
-              {characterData?.result.accessPage.map((page, index) => <S.DescriptionContent key={index}>{page}</S.DescriptionContent>)}
-            </S.DescriptionItem>
-          </S.DescriptionContainer>
-
-          <S.ScenarioContainer>
-            <S.SubTitle>Scenario</S.SubTitle>
-            <S.MainScenarioWrapper>
-              {parsedData.map((item: TItem) => (
-                <S.ScenarioDescription key={item.step}>
-                  {item.step}. {item.actionDescription}
-                </S.ScenarioDescription>
-              ))}
-            </S.MainScenarioWrapper>
-          </S.ScenarioContainer>
-
-          <S.ButtonContainer>
-            <Button color="mint" onClick={() => setModalStep(1)}>
-              Edit and Request
-            </Button>
-            <Button color="blue" onClick={() => dispatch(closeModal())}>
-              Save
-            </Button>
-          </S.ButtonContainer>
-        </S.ConfirmModalContainer>
-      )}
+      {modalStep == 2 && <ScenarioModalStep2 characterData={characterData} setModalStep={setModalStep} />}
     </Modal>
   );
 }
