+++ src/components/scenario/scenarioModal/scenarioModal/scenarioModalStep.style.ts
