+++ src/utils/getPageNumbers.ts
@@ -7,13 +7,9 @@ export const getPageNumbers = ({ totalPages, currentPage }: TGetPageNumbersProps
   if (totalPages === 0) {
     return [0];
   }
-
-  if (totalPages < 5 && totalPages > 0) {
-    return Array.from({ length: totalPages }, (_, index) => index);
-  }
-
-  const startPage = Math.max(0, Math.min(currentPage - 2, totalPages - 5));
-  const endPage = Math.min(totalPages - 1, startPage + 4);
+  const page = Math.floor((currentPage + 1) / 10);
+  const startPage = page * 10;
+  const endPage = Math.min(page + 9, totalPages - 1);
 
   return Array.from({ length: endPage - startPage + 1 }, (_, index) => startPage + index);
 };
